+++ mvlearn/model_selection/__init__.py
@@ -1,3 +1,4 @@
+from .split import train_test_split
 from .validation import cross_validate
 
-__all__ = ["cross_validate", ]
+__all__ = ["train_test_split", "cross_validate", ]